+++ app/src/main/java/com/lumhue/karskrin/lumhue/model/Rgb.java
@@ -2,8 +2,8 @@ package com.lumhue.karskrin.lumhue.model;
 
 public class Rgb {
 
-    public Integer r;
-    public Integer g;
-    public Integer b;
+    public Double r;
+    public Double g;
+    public Double b;
 
 }
